==> 2_descr_cleanup.ipynb <==
# MyGamePass #
## Description and Tag Preprocessing ##

A combined dataset from Steam of over 27,000 games with extensive descriptive information collected from the Steam store and Steam Spy API, packaged and published to Kaggle by Nik Davis. 

https://www.kaggle.com/nikdavis/steam-store-games

Three files will be used from this dataset.  This notebook will focus on the steam_description_data.csv and steamspy_tag_data.csv

The steam.csv master file contains the appid and store information for each game, and it will be merged with the description for each game found in the steam_description_data.csv file.  

Additionally, the steamspy_tag dataframe which includes the counts of each key term for each game will be used for content filtering.

# Import standard packages
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns

# Read in the description dataframe and take a look
descr_df = pd.read_csv('data/steam_store/steam_description_data.csv')
descr_df.head(3)

# Compare the lengths of each description column
appid=30
print('detailed len:',int(descr_df[descr_df['steam_appid']==appid]['detailed_description'].str.len()))
print('about len:',int(descr_df[descr_df['steam_appid']==appid]['about_the_game'].str.len()))
print('short len:',int(descr_df[descr_df['steam_appid']==appid]['short_description'].str.len()))

- As we will be using Natural Language Processing, the more detailed description the better.  Keep detailed and drop the others

# Drop the redundant columns
descr_df.drop(['about_the_game','short_description'],axis=1,inplace=True)
# Rename the remaining columns
descr_df.rename(columns={'steam_appid':'appid','detailed_description':'description'},inplace=True)

descr_df.head(3)

- We need a new dataframe of the high-level video game information
- add appid, name, percent_positive_ratings, and total_ratings, average_playtime from the steam master table

# Load in the master dataframe, only the columns we need
steam_df = pd.read_csv('data/steam_clean.csv',usecols=['appid','name','percent_positive_ratings','total_ratings','average_playtime'])
steam_df.head(3)

# Create a new games dataframe by merging the description
games_df = steam_df.merge(descr_df,on='appid')
games_df.head(3)

# Output to a new csv file and read in to confirm
games_df.to_csv('data/games.csv',index=False)
newdf = pd.read_csv('data/games.csv')
newdf.head(3)

newdf[newdf['name'].str.contains('Mass Effect')]

### Secondary Games Dataframe with additional non-numeric features from Steam master table combined ###

One strategy is to combine the non-numeric features into the description prior to Natural Language Processing so they will be included in the analysis.  We will set that alternative games dataframe up now./

# Read in the master dataframe with the chosen features
df = pd.read_csv('data/steam_store/steam.csv',usecols=['appid','developer','publisher','categories','steamspy_tags'])
df.head(3)

# Combine the features into one column with a separator 
df['combo'] = df['developer'] + ';' + df['publisher'] + ';' + df['categories'] + ';' + df['steamspy_tags']
df.head(3)

# Filter down to the two columns we need
df = df.filter(['appid','combo'])
df.head(3)

# Separate into individual words (punctuation is fine, it will be handled later)
df['combo'] = df['combo'].str.split(';')
df.head()

# Output this to a file
df.to_csv('data/steam_combo.csv',index=False)
newdf = pd.read_csv('data/steam_combo.csv')
newdf.head(3)

# We will be using that games_df from before, let's confirm it is ready
games_df.head(3)

# Merge the two dataframes to combine all of the text data
combo_df = games_df.merge(df, on='appid')
combo_df.head(3)

# Create a new combined column
combo_df['descr_combo'] = combo_df['description'] + str(combo_df['combo'])
# Drop the relic combo column
combo_df.drop('combo',1,inplace=True)
combo_df.head(3)

# Output to a file and confirm
combo_df.to_csv('data/games_combo.csv',index=False)
new_df = pd.read_csv('data/games_combo.csv')
new_df.head(3)

### Fantastic!  We are ready for content based filtering ###
