==> 4_content_modeling.ipynb <==
# MyGamePass #
## Content Based Filtering ##

As a kid, I used to love finding new favorites by going to the store and seeing which game caught my attention with the name and description, if I recognized the developer or knew I loved that genre of game.  

With natural language processing and machine learning we are able to create a recommender system that factors these descriptive features to find similar games.

We began by performing extensive EDA and preprocessing on the dataframes that will be used in this modeling.  The data cleanup folder contains each of the cleanup notebooks for the dataset.

# Begin by importing the standard packages
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt

# NLP Packages
import string
import nltk

# import TfidfVectorizer 
from sklearn.feature_extraction.text import TfidfVectorizer

# Import cosine similarity
from sklearn.metrics.pairwise import cosine_similarity

# Import the prepared games dataset, with the combined description and non-numeric features
games_df = pd.read_csv('data/games_combo.csv')
games_df.head(3)

column: description

- appid: steam appid for joining
- name: video game name
- average_playtime: average playtime based on master steam store dataset
- total_ratings: total number of ratings
- percent_positive_ratings: percentage of ratings that are positive
- description: text description of the game
- descr_combo: combined non-numeric features with the description for each game

# Let's look at the sample output for one of my favorite game franchises
games_df[games_df['name'].str.contains('Halo',na=False)]

- Great.  This is excellent data for the NLP to build the similarity matrix.
- Note: there appear to be some HTML tags that will need to be addressed

## User Independent System ##

Let's start just by exploring the ratings.  A common approach for a "default" recommendation or for a new user would be to simply recommend the highest rated video game in the dataset.

# Top 10 highest rated games, no filtering
top_rated = games_df.sort_values(by=['percent_positive_ratings'],ascending=False)
top_rated.head(10)

Huh...I've never heard of any of these games...

- Right off the bat it is a red flag for any game to have 100% positive ratings.  Humans can be very tough critics and rarely agree uniformally especially on entertainment preferences.

- We will combat this with a simple filter, to only include games above a minimum number of votes
    - This does have the potential to ignore any brand new games, which is a feature I would like to explore to add robustness in a future update

# Add a vote threshold for a more realistic Top 10
# Let's be pretty tough, filtering for very popular games
threshold = 10000
# Temporary dataframe for the games with a lot of ratings
many_ratings_df = games_df[games_df['total_ratings'] >= threshold]
# Sort by most positive
top_rated_v2 = many_ratings_df.sort_values(by=['percent_positive_ratings'],ascending=False)
# Rank the top 10
top_rated_v2.head(10)

- Much better.  And I agree, #1 Portal 2 and #10 The Witcher 3 are two of my favorite games of all time.

- These ratings filters can be used to refine the content recommendations in a similar fashion

Let's take a look at one of the most popular entertainment franchises of all time, Star Wars.  There have been a wide variety of Star Wars games made of various genres, types, with various modes, etc. and the franchise has a dedicated fanbase.  

There are so many Star Wars games it can be hard to find the ones you like, and too many options to play them all.  A recommender system is the solution!

# Star Wars games
games_df[games_df['name'].str.contains('STAR WARS')]

## Natural Language Processing ##

Using the Natural Language Toolkit (NLTK) package we will process the descriptive data into numeric values for our cosine similarity matrix.

descr_combo: the combined description and non-numeric features (developer, publisher, categories, tags)

- TF IDF vectorization of descr_combo
    - TF IDF was chosen as it puts an increased weight on the most unique words found in each description
    - This is preferred as these are the most "descriptive" words as opposed to the most frequent word count
- compared to count vectorization, this method will put more emphasis on the unique words for each description, and less emphasis on the words found throughout all descriptions

# Create a custom function to remove the html tags from the descriptions
# This will be called by the tokenizer first
def remove_html_tags(text):
    """Remove html tags from a string"""
    import re
    clean = re.compile('<.*?>')
    return re.sub(clean, '', text)

# We will use the nltk for our stemming and stopwords in the custome tokenizer function
# initialize the porter stemmer from nltk
# A stemmer will find the root word for a more robust approach
stemmer = nltk.stem.PorterStemmer()
# Download our stopwords 
nltk.download('stopwords')
from nltk.corpus import stopwords
# Use english stopwords - "throwaway" words that add no value for our matrix
stopwords = stopwords.words('english')

# Custom tokenizer to remove html tags, punctuation, set to lowercase, and remove stopwords
def my_tokenizer(sentence):
    # Remove HTML tags with custom function
    sentence = remove_html_tags(sentence)
    
    # remove punctuation using string attribute
    for punct in string.punctuation:
        # set to lower case with built in functions
        sentence = sentence.replace(punct,'').lower()

    # split into words
    words = sentence.split(' ')
    stemmed_list = []
    
    # remove stopwords and any tokens that are just empty strings
    for word in words:
        if (not word in stopwords) and (word!=''):
            # Stem words
            stemmed = stemmer.stem(word)
            stemmed_list.append(stemmed)

    return stemmed_list

%%time
# Set up our TF IDF vectorizer
# Initial thresholds
minimum_descr_count = 5 # do not count words unless they occur in at least this many descriptions
maximum_descr_perc = 0.90 # drop words that occur in 90% or more of the descriptions

# Initialize vectorizer with our stop words, thresholds, and custom tokenizer
vectorizer = TfidfVectorizer(stop_words = stopwords, min_df=minimum_descr_count, 
                             max_df=maximum_descr_perc, tokenizer=my_tokenizer)

# Prepare the descr_combo column for processing
games_df['descr_combo'] = games_df['descr_combo'].fillna('')

# Create the TF_IDF matrix of the prepared descr_combo column
TF_IDF_matrix = vectorizer.fit_transform(games_df['descr_combo'])

# Check the shape (This will be a matrix of every game and every descriptive word, may be quite large)
TF_IDF_matrix.shape

- Great.  We now have our TF_IDF matrix of the descriptive features.  18,207 rows (games) and 15,950 features

# Let's look again at the Halo games in the dataset from before
games_df[games_df['name'].str.contains('Halo',na=False)]

- Now that the game descriptive information is numeric a TF_IDF matrix, we can easily compare the similarities of each game using the cosine_similarity function
- Cosine similarity will compare each game and assign a score from 0 (not at all similar) to 1.0 (100% similar)

# For example, let's compare two of the halo games listed above
# Compare the values from the TF_IDF_matrix for each game
game_1 = TF_IDF_matrix[(games_df['name']=='Halo: Spartan Assault').values,]
game_2 = TF_IDF_matrix[(games_df['name']=='Halo: Spartan Strike').values,]

print("Similarity:", cosine_similarity(game_1,game_2))

- 68.6% is a very high score for a straightforward content-based filtering approach
- Having played both of these games, I can confirm they are extremely similar as one is nearly a copy and paste sequel to the other

In order to create a recommender system from this data we will now need to compare the similarities of every individual game to every other game in the dataset.  This is where the power of machine learning is great as this would have been an extraordinarily time consuming exercise to do by hand (and imagine applying that approach to all 1.2 million games available today!)

cosine_similarity is and extremely powerful function, as it can perform this analysis on our TF_IDF_matrix and create a similarity matrix used to pull recommendations from if they meet our criteria.

%%time
# We compare the similarities of every game to every other game.
similarities = cosine_similarity(TF_IDF_matrix, dense_output=False)

- Wow!  All of that in 14 seconds.  That's amazing!

# Let's test it out with a very popular and influential game, Grand Theft Auto III
games_df[games_df['name']=='Grand Theft Auto III']

# Grab the index of the game to compare to
game_index = games_df[games_df['name']=='Grand Theft Auto III'].index

# Create a dataframe with game appid, name, and similarity
sim_df = pd.DataFrame({'appid':games_df['appid'],'game':games_df['name'],'similarity':np.array(similarities[game_index,:].todense()).squeeze()})

# Sort the dataframe and limit the results to the top 10 most similar games from the matrix
sim_df.sort_values(by='similarity',ascending=False).head(10)

- Somewhat interesting results.  As expected, the other entries in the franchise appear on the list.
- However I would expect more similar games such as Saints Row, or possibly another Rockstar openworld game such as Red Dead Redemption to appear
- In the future this model could be improved by putting more weights on certain factors.
    - As you may recall, we simply included the Developer, Publisher, and other information such as Genre and Categories with the description text.  
    - This likely does not put enough emphasis on these features
    - That could explain why another popular Rockstar game with similar features such as Red Dead Redemption does not appear on this list.
    - Although the game itself is similar (open world, action/adventure, made by the same Developer/Publisher, etc) the descriptions are much different (Wild West vs Urban Crime)
    
This is a great starting point, with many opportunities to improve.  To refine the recommendaions we will add thresholds for the rating information.

As we saw before, it is important to set a minimum vote threshold.  In addition, a minimum rating will be considered.  

These settings can be tuned for each user, as some gamers may prefer only the most popular and highest rated games, whereas others may be more likely to explore to find a lesser-known hit.

# Add a function to restrict results above a total voter and rating percentage threshold 
def content_recommender(name, games, similarities, vote_threshold=1000, rating_threshold=0.7) :
    
    # Get the game by the title
    game_index = games[games['name']==name].index
    
    # Create a dataframe with the game id, name, and rating information with similarity
    sim_df = pd.DataFrame(
        {'appid': games['appid'],
         'game': games['name'], 
         'similarity': np.array(similarities[game_index, :].todense()).squeeze(),
         'vote_count': games['total_ratings'],
         'percent_positive_ratings': games['percent_positive_ratings']
        })
    
    # Get the top 10 games that satisfy our thresholds
    top_games = sim_df[(sim_df['vote_count']>vote_threshold) & 
                       (sim_df['percent_positive_ratings']>rating_threshold)].sort_values(by='similarity', ascending=False).head(10)
    
    return top_games

Excellent!  We now have a working content-based recommender system with adjustable thresholds to customize for each user.

Let's test it out below with some popular games and see how it does now.

# Test the recommender
# Show the 10 most similar games above the threshold, sorted by rating
# Increase the rating threshold to 80% positive reviews
similar_games = content_recommender('Grand Theft Auto III', games_df, similarities, 
                                    rating_threshold=0.80)
similar_games.head(5).sort_values(by='percent_positive_ratings',ascending=False)

- As we can see, the results are much different from before.  Now, more emphasis is being placed on finding "good" games that are also similar as opposed to descriptive similarity alone

# Mass Effect Recommendation
similar_games2 = content_recommender('Mass Effect', games_df, similarities, 
                                     vote_threshold=1000, rating_threshold=0.80)
similar_games2.head(5).sort_values(by='percent_positive_ratings',ascending=False)

- As expected, the sequel to the great Mass Effect is the top recommended game.
- Star Wars: Republic Commando looks to be another great option to check out as well.
    - Another Sci-Fi, action game with futuristic guns and vehicles, that is a pretty solid recommendation
    
Which reminds me of our Star Wars example from before, with so many different games in the franchise it can be hard to pick which one to try next.  I know one of my favorites of all time was Star Wars: Knights of the Old Republic, let's try that.

# Star Wars: Knights of the Old Republic Recommendation
similar_games = content_recommender('STAR WARS™ - Knights of the Old Republic™', games_df, similarities, 
                                     vote_threshold=1000, rating_threshold=0.80)
similar_games.head(5).sort_values(by='percent_positive_ratings',ascending=False)

- Great!  Not surprised to see only Star Wars games on the list, but this is a great way to find a new favorite.  
- I do love Jedi games with lightsabers and force powers, I will have to check out Jedi Knight!

The next notebook will utilize user data for collaborative filtering.

Additionally, further work could be done to properly emphasize the features with custom weights as mentioned previously.

Topic clustering can be performed as well in an effort to enhance the "cold user" approach where someone has never played any games before, but would like a more personalized recommendation on where to start.

A hybrid model utilizing a blend of user-item recommendations and content filtering would be the best approach.  You could find the top games based on similar profiles, and additionally look for more games like those to expand the recommendations further.

# MyGamePass #
## Ben Polzin ##
