==> 5_matrix_factorization.ipynb <==
# MyGamePass #
## Collaborative Based Filtering ##

One of the most compelling ways I would discover new games as a kid was when my best friends would recommend them to me.  My cousin changed my life when he loaned me The Legend of Zelda: Ocarina of Time as a kid and I fell in love.

This experience can be taken to a whole new level with machine learning.  With extensive user engagement data, we can perform analysis to find similar user profiles as your own to predict what ratings you would give a game you've not played before.  

Using these predicted ratings that satisfy our requirements recommendations are provided based on playing habits.

## Matrix Factorization ##

One of the major challenges in putting recommender system applications into production is the nearly impossible compute requirements of an in-memory system when deployed at scale.  Consider if we had 10s of millions of users and millions of games, the standard memory based collaborative filtering would become very challenging.

The solution for this is Matrix Factorization.  We will be using the Funk Singular Vector Decomposition method, or FunkSVD named after the inventor Simon Funk.  It is a part of the Surprise package which has become very popular for its built-in recommender system evaluation tools.

The SVD will generate a predicted rating for every user-item pair in our prepared dataset.

# Import the trinity
import numpy as np 
import pandas as pd 
import matplotlib.pyplot as plt

# Import the surprise packages
from surprise import Dataset
from surprise.reader import Reader
from surprise.prediction_algorithms.matrix_factorization import SVD as FunkSVD

# Import GridSearchCV for algorithm tuning
from surprise.model_selection import GridSearchCV

# Import train_test_split
from surprise.model_selection import train_test_split

Again to highlight, additional extensive data cleaning and preprocessing steps were performed in the data cleanup notebooks.  The prepared data will be loaded from here.

# Read in the prepared dataframe from the user_cleanup notebook
user_df = pd.read_csv('data/user_clean.csv')
user_df.head()

column: description

- user_id: unique id for each user
- app_name: name of video game
- hours_played: amount of hours this user played the video game
- appid: steam appid for joining
- user_num_games: total number of games the user has played in this dataset
- game_popularity: total number of users who have played this game in the dataset
- rating_5: converted the hours_played to a 1-5 rating for Surprise package modeling
- rating_10: converted the hours_played to a 1-10 rating for Surprise package modeling

two ratings were prepared to test impact on accuracy of the model

# And read in the prepared games dataframe from the steam
games_df = pd.read_csv('data/games.csv')
games_df.head()

column: description

- appid: steam appid for joining
- name: video game name
- average_playtime: average playtime based on master steam store dataset
- total_ratings: total number of ratings
- percent_positive_ratings: percentage of ratings that are positive
- description: text description of the game

# Merge the two dataframes on appid
df = user_df.merge(games_df,on='appid')
df = df.drop('name',1)
df.head()

# Let's take a look at one of the most prominent users in the dataset, user 24469287
df[df['user_id'] == 24469287]

# Let's find this users favorite games using the 1-5 rating scale
print(f"Shape:{df[(df['user_id'] == 24469287) & (df['rating_5'] == 5)].shape}")
display(df[(df['user_id'] == 24469287) & (df['rating_5'] == 5)])

- Looking at this list, although there are a lot of great games here it feels like there needs to be more granularity in this dataset
- For example, playing a game 23 hours is vastly different than playing a game 119 hours
- Let's see if the model agrees, however.  More granularity may just be more noise to overfit

# Let's find this users favorite games based on the 1-10 rating scale
print(f"Shape:{df[(df['user_id'] == 24469287) & (df['rating_10'] == 10)].shape}")
display(df[(df['user_id'] == 24469287) & (df['rating_10'] == 10)])

- This game did stand out from the list before, and with the more detailed scale it shows as the only 10 rating

# Prepare the dataframes for the surprise package
# Dataframe needs to contain 3 columns: user id, item id, and rating
# For the 1-10 scale
rating_10_df = df.filter(['user_id','appid','rating_10'])
rating_10_df = rating_10_df.sort_values(by=['user_id','appid'])
# And the 1-5 scale
rating_5_df = df.filter(['user_id','appid','rating_5'])
rating_5_df = rating_5_df.sort_values(by=['user_id','appid'])

## Funk SVD ##
- I will try it first with the 1-10 ratings, it seemed like the added detail would be better

# Confirm dataframe is set up properly (user, item, rating)
rating_10_df.head()

# initialize the reader with 1-10 rating scale
my_reader = Reader(rating_scale=(0,10))

# load the dataframe with the reader
my_dataset_10 = Dataset.load_from_df(rating_10_df, my_reader)

%%time

# Set the parameter grid for optimization
param_grid = {
    # Number of latent factors. More factors could give better results, but can also lead overfitting
    'n_factors': [50, 100, 150], 
    # Number of epochs. Number of iterations the algorithm will run
    'n_epochs': [10, 20, 50],
    # Learning rate. The speed at which algorithm learns. Larger values give faster learning, but smaller values give more accurate learning.
    'lr_all': [0.005, 0.1],
    'biased': [False] }

# Set GridSearchCV with 5 fold cross-validation using the FunkSVD
GS = GridSearchCV(FunkSVD, param_grid, measures=['rmse','mae','fcp'], cv=5)

# Fit the model to the data
GS.fit(my_dataset_10)

# Print the evaluation metrics
print('Root Mean Squared Error (RMSE):',GS.best_score['rmse'])
print('Mean Absolute Error (MAE):',GS.best_score['mae'])
print('Fraction of Concordant Pairs (FCP):',GS.best_score['fcp'])

- Root Mean Squared Error and Mean Absolute Error are measures of error between the predicted ratings and the actual ratings
    - The lower the number the better
- Fraction of Concordant Pairs describe the relationship between the pairs.  A pair is concordant if one user gave a game a high rating and the predicted rating for that game is also high, or vice vera
    - FCP ranges from 0.0 on the low end to 1.0 on the high end
    - High FCP means most of the pairs (predicted ratings and actual ratings) agree in overall sentiment
- This is something to work with.  Humans are finicky, especially with entertainment preferences.  Not great, but not terrible.

Looking at these metrics, the FCP score is preferred, but let's check the best parameters of the two measures

GS.best_params['rmse']

GS.best_params['fcp']

We will optimize on the fcp from here

- let's compare this fcp score to the 0-5 rating scale data

rating_5_df.head()

# Set the reader with accurate rating scale
my_reader = Reader() # default 1-5 rating scale

# Set the dataset
# Remember that the df parameter has to have 3 columns:
# User ids, Item ids (game appid), Ratings
my_dataset_5 = Dataset.load_from_df(rating_5_df, my_reader)

%%time
GS.fit(my_dataset_5)

GS.best_score['fcp']

GS.best_params['fcp']

Using the same gridsearch with the 1-5 rating dataset caused a slight increase in the best FCP score.  Additionally, the best parameters for that score were different for each input.

Let's focus on the 1-5 rating scale as it is more accurate.  This is likely due to an increase in noise for the 1-10 rating, the more simplified model led to more accurate predictions.

%%time

# Expand the parameter grid to really fine tune (this will take longer)
param_grid = {
    # Number of latent factors. More factors could give better results, but can also lead overfitting
    'n_factors': [50, 100, 150, 200, 250], 
    # Number of epochs. Number of iterations the algorithm will run
    'n_epochs': [10, 20, 50, 75, 100],
    # Learning rate. The speed at which algorithm learns. Larger values give faster learning, but smaller values give more accurate learning.
    'lr_all': [0.005, 0.01, 0.1, 1, 10],
    'biased': [False] }

# Set GridSearchCV with 5 fold cross validation (focusing on the 'fcp' measure)
GS2 = GridSearchCV(FunkSVD, param_grid, measures=['fcp'], cv=5)

# Fit the model
GS2.fit(my_dataset_5)

# Check the score - previous best was 0.6257
print('Fraction of Concordant Pairs (FCP):',GS2.best_score['fcp'])

- Nice! An increase to 0.6318.  I'll take it.

# Check best params - previous best was 50 factors, 20 epochs, 0.005 learning rate
GS2.best_params['fcp']

- Looks like the hyperparameters indeed did get optimized with the expanded gridsearch.  We will use these settings from here on out

## Making Predictions with Collaborative Filtering ##

Now that we have our hyperparameters optimized, let's build our collaborative recommender system.

We will use a train_test_split to confirm the accuracy of the predictions and look closer at the performance by extracting the predictions into a dataframe to explore

# Split the data into the train set and the test set, reserving 25% of the data for the testset
trainset, testset = train_test_split(my_dataset_5, test_size=0.25)

# Initialize the FunkSVD
my_svd = FunkSVD(n_factors=200, 
                 n_epochs=100, 
                 lr_all=0.01, 
                 biased=False,
                 verbose=0)
# Fit train set
my_svd.fit(trainset)

# Test the predictions on the testset
my_pred = my_svd.test(testset)

# Put my_pred result in a dataframe
df_prediction = pd.DataFrame(my_pred, columns=['user_id',
                                                     'appid',
                                                     'actual_rating',
                                                     'predicted_rating',
                                                     'details'])

# Calculate the difference of actual and prediction into diff column
df_prediction['diff'] = abs(df_prediction['predicted_rating'] - 
                            df_prediction['actual_rating'])

df_prediction.head()

df_prediction[df_prediction['diff'] <= 1]['user_id'].count() / df_prediction['user_id'].count()

- 50.9% of the dataset the predicted rating is within 1 point of the acutal rating.  Not bad.

Now let's continue with the full trainset and full testset

%%time
# Build full trainset
full_trainset = my_dataset_5.build_full_trainset()

# Re-Fit the FunkSVD from before with the full_trainset
my_svd.fit(full_trainset)

# Define the full test set, fill the empty ratings with -1 where the user has not played the game
full_testset = full_trainset.build_anti_testset(fill=-1)

# Make the predictions on the full_testset
my_prediction = my_svd.test(full_testset)

# Put my_pred result in a dataframe like before
df_prediction = pd.DataFrame(my_pred, columns=['user_id',
                                                     'appid',
                                                     'actual_rating',
                                                     'predicted_rating',
                                                     'details'])

# Check our favorite user id `24469287` for the top predictions
newdf = df_prediction[df_prediction['user_id'] == 24469287].sort_values(
    by=['predicted_rating'], ascending=False).head()

# Merge the dataframes
merge_df = newdf.merge(games_df, how='left', 
                    left_on=['appid'], right_on=['appid'])

# Show the top recommendations
merge_df

You could take this collaborative modeling even further with a hybrid model as mentioned at the end of the content modeling notebook.  A hybrid model could function by finding the top predictions based on similar profiles, and then also looking for additional games that are similar to those based on their content to expand the recommendations.

Additionally, with much more detailed user data you could perform extensive analysis to better define a similar user profile.  For example, someone who only plays on the weekends may prefer different games than someone who regularly plays after work every day, even if the total hours played appears to be similar.  Of course, demographic data would be valuable as well.  

A particularly interesting analysis I would like to perform would be to cross-reference a users Netflix profile as you can learn a lot about a users entertainment preferences and cross-over between the two medium.

# MyGamePass #
## Ben Polzin ##
